move simulation setup to respective submodule
