+++ MainLPM.jl
@@ -11,20 +11,10 @@ using SocialABMs: MultiABM
 using LoneParentsModel.Loaders:    loadUKMapParameters, loadUKPopulationParameters
 using LoneParentsModel.Declare:    createUKDemography 
 using LoneParentsModel.Initialize: initializeDemography!
-using LoneParentsModel.SimSetup:   loadSimulationParameters
+using LoneParentsModel.SimSetup:   loadSimulationParameters, LPMUKDemography
 
-using SocialABMs: dummystep 
-
-using SocialSimulations: ABMSocialSimulation, MABMSimulation
-using SocialSimulations: run!, attach_agent_step!, attach_model_step!  
-
-import SocialSimulations: setup!, AbstractExample
-
-# export setup!, LPMUKDemography 
-
-# Example Name 
-
-struct LPMUKDemography <: AbstractExample end 
+using SocialSimulations: MABMSimulation
+using SocialSimulations: run! 
 
 # Model parameters 
 
@@ -54,24 +44,10 @@ println(); println();
 
 simProperties = loadSimulationParameters()
 
-function setup!(sim::ABMSocialSimulation,example::LPMUKDemography) 
-    attach_agent_step!(sim,dummystep) 
-    attach_model_step!(sim,dummystep)
-    nothing 
-end
-
-function setup!(sim::MABMSimulation,example::LPMUKDemography) 
-    #= 
-    n = length(sim.simulations) 
-    for i in 1:n 
-        attach_agent_step!(simulations[i],X) 
-        attach_model_step!(simulations[i],Y)
-    end
-    =# 
-    nothing 
-end
-
 lpmDemographySim = MABMSimulation(ukDemography,simProperties, 
                                   example=LPMUKDemography())
 
+# Execution 
+
 run!(lpmDemographySim,verbose=true)
+
